# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Absen Ok API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/task/task-list": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Project Manager & Member Access - Get All Task",
                "description": "The user retrieves all task data",
                "operationId": "App\\Http\\Controllers\\Dashboard\\Task\\TaskController::getTaskList",
                "responses": {
                    "200": {
                        "description": "Success: All Task retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Get All Task.",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Get All Task.",
                                            "type": "string",
                                            "example": "All Task retrieved successfully."
                                        },
                                        "data": {
                                            "description": "Get All Task.",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "description": "Get All Task.",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "description": "Get All Task.",
                                                        "type": "string",
                                                        "example": "Tugas Pertama pada Project Pertama"
                                                    },
                                                    "project_id": {
                                                        "description": "Get All Task.",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "deadline": {
                                                        "description": "Get All Task.",
                                                        "type": "string",
                                                        "example": "2023-11-27 10:50:10"
                                                    },
                                                    "description": {
                                                        "description": "Get All Task.",
                                                        "type": "text",
                                                        "example": "Ini adalah tugas pertama saya dari Project pertama"
                                                    },
                                                    "status": {
                                                        "description": "Get All Task.",
                                                        "type": "string",
                                                        "enum": [
                                                            "to-do",
                                                            "in progress",
                                                            "completed"
                                                        ],
                                                        "example": "completed"
                                                    },
                                                    "comment": {
                                                        "description": "Get All Task.",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "comment": {
                                                                    "description": "Get All Task.",
                                                                    "type": "text",
                                                                    "example": "Ini adalah komentar 1"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Approved Tasks"
                ],
                "summary": "Project Manager Access - edit the status in the approved task table",
                "description": "The user retrieves all task data",
                "responses": {
                    "200": {
                        "description": "Success: status changed successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "edit the status in the approved task table.",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "edit the status in the approved task table.",
                                            "type": "string",
                                            "example": "status changed successfully."
                                        },
                                        "data": {
                                            "description": "edit the status in the approved task table.",
                                            "properties": {
                                                "user_id": {
                                                    "description": "edit the status in the approved task table.",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "task_id": {
                                                    "description": "edit the status in the approved task table.",
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "status": {
                                                    "description": "edit the status in the approved task table.",
                                                    "type": "string",
                                                    "example": "approved"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/task/{task_id}/list-comment": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Project Manager & Member Access - Get comments for a specific task",
                "description": "Users take comment based on Projects",
                "responses": {
                    "200": {
                        "description": "Success: comments based on the task being successfully retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Get comments for a specific task.",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Get comments for a specific task.",
                                            "type": "string",
                                            "example": "comments based on the task being successfully retrieved."
                                        },
                                        "tasks": {
                                            "description": "Get comments for a specific task.",
                                            "properties": {
                                                "comment": {
                                                    "description": "Get comments for a specific task.",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "comment": {
                                                                "description": "Get comments for a specific task.",
                                                                "type": "text",
                                                                "example": "Ini adalah komentar 1"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: task not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Get comments for a specific task.",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Get comments for a specific task.",
                                            "type": "string",
                                            "example": "task not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}